+++ R/00-NAMESPACE.R
@@ -5,23 +5,23 @@
 #' @importFrom coda raftery.diag gelman.diag geweke.diag heidel.diag mcmc
 #' @importFrom DataTools downloadModelUI downloadModelServer importDataUI importDataServer
 #'  tryCatchWithWarningsAndErrors
-#' @importFrom dplyr ungroup summarise group_by "%>%" summarise_ group_by_
+#' @importFrom dplyr "%>%" bind_cols bind_rows group_by group_by_ left_join summarise summarise_
+#'  ungroup
 #' @importFrom DT datatable renderDT DTOutput
 #' @importFrom ggplot2 aes_string element_text geom_errorbar geom_point ggplot
 #'  labs scale_x_discrete theme aes aes_ geom_boxplot ylab xlab
 #' @importFrom graphics plot boxplot
 #' @importFrom grDevices dev.off pdf png svg tiff
-#' @importFrom jsonlite toJSON
-#' @importFrom openxlsx write.xlsx
 #' @importFrom Rcpp evalCpp
 #' @importFrom pROC plot.roc
 #' @importFrom rstan extract
 #' @importFrom shinyjs useShinyjs alert
+#' @importFrom shinyTools dataExportButton dataExportServer
 #' @importFrom shinyWidgets pickerInput updatePickerInput
-#' @importFrom stats lm as.formula predict rexp rnorm terms rpois median quantile sd qnorm 
-#'  chisq.test cor lm.fit model.matrix na.omit
+#' @importFrom stats as.formula chisq.test cor lm lm.fit median model.matrix na.omit 
+#'  predict qnorm quantile rexp rnorm rpois sd setNames terms
 #' @importFrom tidyr gather
-#' @importFrom utils capture.output head packageVersion read.csv write.table
+#' @importFrom utils capture.output head packageVersion
 #' @importFrom yaml read_yaml
 NULL