+++ R/02-modelDiagnostics.R
@@ -5,13 +5,17 @@ modelDiagnosticsTab <- function(id) {
     "Model Diagnostics",
     value = "summaryTab",
     selectInput(ns("modelSelection"), "Select model", choices = ""),
-    radioButtons(ns("diagType"), label = "Diagnostics Type", choices = c("Gelman Scale Reduction Factor" = "gelman",
-                                                                         "Raftery and Lewis" = "raftery",
-                                                                         "Geweke z-Score" = "geweke",
-                                                                         "Heidelberger-Welch" = "heidel")),
+    radioButtons(ns("diagType"), label = "Diagnostics Type", choices = getDiagnosticTypes()),
     verbatimTextOutput(ns("diagnostics")),
     textExportButton(ns("exportText"))
   )
+}
+
+getDiagnosticTypes <- function() {
+  c("Gelman Scale Reduction Factor" = "gelman",
+    "Raftery and Lewis" = "raftery",
+    "Geweke z-Score" = "geweke",
+    "Heidelberger-Welch" = "heidel")
 }
 
 modelDiagnostics <- function(input, output, session, model, nChains) {
@@ -20,18 +24,26 @@ modelDiagnostics <- function(input, output, session, model, nChains) {
     updateSelectInput(session, "modelSelection", choices = names(model()$models))
   })
   
+  allDiagnostics <- reactiveVal()
+  observe({
+    req(model())
+    
+    thisDiagnostics <- extractAllDiagnostics(allModels = model()$models,
+                                             nChains = nChains) %>% 
+      bindAllResults(addEmptyRow = TRUE)
+    
+    allDiagnostics(thisDiagnostics)
+  })
+  
   printFun <- reactive({
     req(model())
     req((input$modelSelection %in% names(model()$models)))
     
     function() {
-      varNames <- model()$models[[input$modelSelection]]@varNames
-      parameters <- extract(model()$models[[input$modelSelection]])$betaAll
-      colnames(parameters) <- varNames
-      parameters <- as.data.frame(parameters)
-      diag <- convergenceDiagnostics(parameters, nChains)
-      diagType <- input$diagType
-      print(diag[[diagType]])
+      printDiagnostics(allModels = model()$models,
+                       modelName = input$modelSelection,
+                       nChains = nChains,
+                       diagType = input$diagType)
     }
   })
   
@@ -41,6 +53,75 @@ modelDiagnostics <- function(input, output, session, model, nChains) {
   })
   
   callModule(textExport, "exportText", printFun = printFun, filename = "diagnostics")
+  
+  return(allDiagnostics)
+}
+
+extractAllDiagnostics <- function(allModels, nChains, asDataFrame = TRUE) {
+  modelNames <- names(allModels)
+  names(modelNames) <- modelNames
+  
+  lapply(modelNames, function(x) {
+    # one model per resRow
+    allDiagTypes <- names(getDiagnosticTypes())
+    names(allDiagTypes) <- names(getDiagnosticTypes())
+    resRow <- lapply(allDiagTypes, function(type) {
+      # one diagnostic per resCol
+      resCol <- capture.output({
+        printDiagnostics(allModels = allModels,
+                         modelName = x,
+                         nChains = nChains,
+                         diagType = getDiagnosticTypes()[type])
+      })
+      if (asDataFrame) {
+        # remove first row if empty
+        if (gsub(" ", "", resCol[1]) == "") resCol <- resCol[2:length(resCol)]
+        
+        # convert to dataframe
+        resCol <- resCol %>%
+          as.data.frame()
+        colnames(resCol) <- type
+      }
+      
+      resCol
+    })
+    
+    if (asDataFrame) {
+      # extend shorter columns to length of longer columns
+      maxNRows <- max(sapply(resRow, nrow))
+      
+      resRow <- lapply(names(resRow), function(type) {
+        resCol <- resRow[[type]]
+        if (nrow(resCol) < maxNRows) {
+          tmp <- resCol[[type]]
+          length(tmp) <- maxNRows
+          tmp <- tmp %>%
+            as.data.frame()
+          colnames(tmp) <- type
+          
+          resCol <- tmp
+        }
+        
+        resCol
+      })
+      
+      # bind columns of diagnostics
+      resRow <- resRow %>%
+        bind_cols() %>%
+        prefixNameAsColumn(name = "Model", value = x)
+    }
+    
+    resRow
+  })
+}
+
+printDiagnostics <- function(allModels, modelName, nChains, diagType) {
+  varNames <- allModels[[modelName]]@varNames
+  parameters <- extract(allModels[[modelName]])$betaAll
+  colnames(parameters) <- varNames
+  parameters <- as.data.frame(parameters)
+  diag <- convergenceDiagnostics(parameters, nChains)
+  print(diag[[diagType]])
 }
 
 convergenceDiagnostics <- function(parameters, nChains){